Move arrival creation to the booking client
