+++ server/services/arrivalService.test.ts
@@ -1,33 +1,33 @@
 import type { Arrival } from 'approved-premises'
 
 import ArrivalService from './arrivalService'
-import ArrivalClient from '../data/arrivalClient'
+import BookingClient from '../data/bookingClient'
 import ArrivalFactory from '../testutils/factories/arrival'
 
-jest.mock('../data/arrivalClient.ts')
+jest.mock('../data/bookingClient.ts')
 
 describe('ArrivalService', () => {
-  const arrivalClient = new ArrivalClient(null) as jest.Mocked<ArrivalClient>
-  const arrivalClientFactory = jest.fn()
+  const bookingClient = new BookingClient(null) as jest.Mocked<BookingClient>
+  const bookingClientFactory = jest.fn()
 
-  const service = new ArrivalService(arrivalClientFactory)
+  const service = new ArrivalService(bookingClientFactory)
 
   beforeEach(() => {
     jest.resetAllMocks()
-    arrivalClientFactory.mockReturnValue(arrivalClient)
+    bookingClientFactory.mockReturnValue(bookingClient)
   })
 
   describe('createArrival', () => {
     it('on success returns the arrival that has been posted', async () => {
       const arrival: Arrival = ArrivalFactory.build()
       const token = 'SOME_TOKEN'
-      arrivalClient.create.mockResolvedValue(arrival)
+      bookingClient.markAsArrived.mockResolvedValue(arrival)
 
       const postedArrival = await service.createArrival(token, 'premisesID', 'bookingId', arrival)
       expect(postedArrival).toEqual(arrival)
 
-      expect(arrivalClientFactory).toHaveBeenCalledWith(token)
-      expect(arrivalClient.create).toHaveBeenCalledWith('premisesID', 'bookingId', arrival)
+      expect(bookingClientFactory).toHaveBeenCalledWith(token)
+      expect(bookingClient.markAsArrived).toHaveBeenCalledWith('premisesID', 'bookingId', arrival)
     })
   })
 })
